% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{buildFilePath}
\alias{buildFilePath}
\title{Helper function to build a file path to the correct output folder using a
specified suffix}
\usage{
buildFilePath(suffix, pep = prj)
}
\arguments{
\item{suffix}{A file suffix identifier}

\item{pep}{A PEP project configuration file}
}
\description{
Helper function to build a file path to the correct output folder using a
specified suffix
}
