% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppro.R
\name{calcFRiF}
\alias{calcFRiF}
\title{Calculate the Fraction of Reads in Features (FRiF)}
\usage{
calcFRiF(bedFile, reads)
}
\arguments{
\item{bedFile}{A BED format file}

\item{reads}{Number of aligned reads}
}
\description{
This function calculates the fraction of reads in a feature and returns
a modified BED file with the cumulative sum of reads, cumulative size
of covered features, the fraction of reads in those features, and the
number of total features.
}
\examples{
calcFRiF()
}
\keyword{FRiF}
