% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppro.R
\name{computeLimit}
\alias{computeLimit}
\title{Compute the axis value limit}
\usage{
computeLimit(value, ccurve_TOTAL_READS)
}
\arguments{
\item{value}{An axis limit value.}

\item{ccurve_TOTAL_READS}{A vector of read counts from a sample.}
}
\description{
This function returns the index of ccurve_TOTAL_READS containing the
closest value to x_max
}
\examples{
computeLimit()
}
\keyword{limit}
\keyword{preseq}
