% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{consensusPeaks}
\alias{consensusPeaks}
\title{This function is meant to identify a project level set of consensus peaks.}
\usage{
consensusPeaks(pep)
}
\arguments{
\item{pep}{A PEP configuration file}
}
\description{
This function is meant to identify a project level set of consensus peaks.
}
\keyword{consensus}
\keyword{peaks}
