% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{countReproduciblePeaks}
\alias{countReproduciblePeaks}
\title{Internal helper function for \code{collapsePeaks}
Count the number of times a peak appears across a list of peak files}
\usage{
countReproduciblePeaks(peak_list, peak_DT)
}
\arguments{
\item{peak_list}{A list of data.table objects representing
narrowPeak files}

\item{peak_DT}{A single data.table of peaks}
}
\description{
Internal helper function for \code{collapsePeaks}
Count the number of times a peak appears across a list of peak files
}
