% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{createAssetsSummary}
\alias{createAssetsSummary}
\title{Create and return assets spreadsheet and save the spreadsheet to file}
\usage{
createAssetsSummary(pep)
}
\arguments{
\item{pep}{A PEP configuration file}
}
\description{
Create and return assets spreadsheet and save the spreadsheet to file
}
