% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{createStatsSummary}
\alias{createStatsSummary}
\title{Create and return sample statistics summary data table}
\usage{
createStatsSummary(samples, results_subdir)
}
\arguments{
\item{samples}{A PEP project character vector of sample names}

\item{results_subdir}{A PEP project results subdirectory path}
}
\description{
Create and return sample statistics summary data table
}
