% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{fileIcon}
\alias{fileIcon}
\title{Create and return a file icon}
\usage{
fileIcon(x = seq(0:275), y = seq(0:275))
}
\arguments{
\item{x}{Range of x-axis dimensions}

\item{y}{Range of y-axis dimensions}
}
\description{
Create and return a file icon
}
\keyword{file}
\keyword{icon}
