% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{filetype}
\alias{filetype}
\title{Assign the filetype to a variable and make sure we close the connection}
\usage{
filetype(path)
}
\arguments{
\item{path}{A path to a file for which you wish to determine its class}
}
\description{
Assign the filetype to a variable and make sure we close the connection
}
