% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{getAbbr}
\alias{getAbbr}
\title{Determine the appropriate abbreviation for large numbers}
\usage{
getAbbr(vec)
}
\arguments{
\item{vec}{A vector of numbers}
}
\description{
Modified From: https://stackoverflow.com/questions/28159936/formatting-large-currency-or-dollar-values-to-millions-billions
}
\keyword{abbreviation}
