% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{getPrealignments}
\alias{getPrealignments}
\title{Return a list of prealignments from a stats_summary.tsv file if they exist}
\usage{
getPrealignments(stats_file)
}
\arguments{
\item{stats_file}{A looper derived stats_summary.tsv file}
}
\description{
Return a list of prealignments from a stats_summary.tsv file if they exist
}
