% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{is.empty}
\alias{is.empty}
\title{Check if a data frame is empty.}
\usage{
is.empty(df)
}
\arguments{
\item{df}{A data frame to check for emptiness}
}
\description{
Return TRUE if provided data frame is NULL, has no rows, or has no columns.
}
