% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{mode}
\alias{mode}
\title{Calculate mode(s) of data}
\usage{
mode(x, return_multiple = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of numbers or characters}

\item{return_multiple}{Bool to return multiple modes or first in order}

\item{na.rm}{Bool Remove NAs}
}
\description{
From: https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
\keyword{mode}
