% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{narrowPeakToBigBed}
\alias{narrowPeakToBigBed}
\title{Convert a narrowPeak file to a bigBED format file.}
\usage{
narrowPeakToBigBed(
  input = input,
  chr_sizes = chr_sizes,
  ucsc_tool = ucsc_tool,
  keep = FALSE
)
}
\arguments{
\item{input}{Path to narrowPeak file}

\item{chr_sizes}{Genome chromosome sizes file. <Chr> <Size>}

\item{ucsc_tool}{Path to UCSC tool "bedToBigBed"}

\item{keep}{Keep BED format intermediate file}
}
\description{
Convert a narrowPeak file to a bigBED format file.
}
\keyword{bigBed}
\keyword{narrowPeak}
