% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\docType{package}
\name{PEPATACr}
\alias{PEPATACr}
\title{PEPPRO R tools (PEPATACr)}
\description{
PEPATACr is a collection of functions to generate QC and analysis plots
for ATAC-seq data analysis.
}
\references{
\url{http://github.com/databio/pepatac}
}
\author{
Jason Smith
}
