% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppro.R
\docType{package}
\name{PEPPROr}
\alias{PEPPROr}
\alias{PEPPROr-package}
\title{PEPPRO R tools (PEPPROr)}
\description{
PEPPROr is a collection of functions to generate QC and analysis plots
for PRO-(or GRO-)seq data analysis.
}
\references{
\url{http://github.com/databio}
}
\author{
Jason Smith
}
