% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotAlignedPct}
\alias{plotAlignedPct}
\title{This function is meant to plot project level percent aligned reads.}
\usage{
plotAlignedPct(prj, summary_dir, stats)
}
\arguments{
\item{prj}{A PEPr Project object}

\item{summary_dir}{A PEP project summary directory path}

\item{stats}{A looper derived stats_summary.tsv file}
}
\description{
This function is meant to plot project level percent aligned reads.
}
\keyword{aligned}
\keyword{percent}
\keyword{reads}
