% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotAnno}
\alias{plotAnno}
\title{Plot read and peak annotations and distributions}
\usage{
plotAnno(
  plot = c("chromosome", "tss", "genomic"),
  input,
  type = c("np", "bed"),
  feat,
  genome = "hg38",
  output = "chromosome_distribution.pdf"
)
}
\arguments{
\item{plot}{Choose the type of plot to produce}

\item{input}{input file to be annotated}

\item{feat}{BED6 file containing annotated features of interest}

\item{genome}{genome identifier}

\item{output}{path to output directory}

\item{peaks}{BED or narrowPeak format file of called peaks}
}
\description{
This function is meant to annotate called peaks or aligned reads by known or
predicted genomic regions
}
\examples{
data("peaks")
data("feat")
plotAnno(feat = "feat", input="peaks", type = "np",
         genome = "hg38", output = ".")
}
\keyword{annotation}
\keyword{peak}
\keyword{read}
