% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotComplexityCurves}
\alias{plotComplexityCurves}
\title{Plot library complexity curves}
\usage{
plotComplexityCurves(
  ccurves,
  coverage = 0,
  read_length = 0,
  real_counts_path = NA,
  ignore_unique = FALSE,
  x_min = 0,
  x_max = 5e+08
)
}
\arguments{
\item{ccurves}{A single preseq output file from one sample.}

\item{coverage}{Use coverage on axes instead of read counts. Enter the number
of base pairs of your reference genome.}

\item{read_length}{Sequence read length, for use in coverage calculations.}

\item{real_counts_path}{File name for a file with three columns -
preseq filename, total number of reads, number of
unique reads
(unique is optional, file is whitespace delimited)}

\item{ignore_unique}{If FALSE, ignore information about unique read counts
found in real_counts_path file.}

\item{x_min}{Lower x-limit (default 0)}

\item{x_max}{Upper x-limit (default 500 million)}

\item{more_ccurves}{A list of additional preseq output files from more samples}
}
\description{
This function plots library complexity curves using data from
preseq and produces pdf/png output files.
Adapted from preseq_complexity_curves.py from
https://github.com/ewels/ngi_visualizations/tree/master/ngi_visualizations/preseq_complexity_curves
}
\examples{
data("ccurve")
data("counts")
plotComplexityCurves(ccurves = "ccurve", coverage=3099922541, read_length=30,
                     real_counts_path="counts")
}
\keyword{complexity}
\keyword{library}
\keyword{preseq}
