% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotFRiF}
\alias{plotFRiF}
\title{Plot Fraction of Reads in Features (FRiF)}
\usage{
plotFRiF(
  sample_name,
  num_reads,
  genome_size,
  type = c("cfrif", "frif", "both"),
  reads = TRUE,
  output_name,
  bedFile
)
}
\arguments{
\item{sample_name}{Name of sample}

\item{num_reads}{Number of aligned reads in sample}

\item{genome_size}{Size of genome in bp}

\item{output_name}{Output file name}

\item{bedFile}{A BED format file}
}
\description{
This function plots the fraction of reads in a set of features
and produces pdf/png output files.
}
\examples{
data("promoter")
data("promoter_flanking")
data("exon")
data("intron")
data("utr3")
data("utr5")
plotFRiF(sample_name="example", num_reads=87520,
         output_name="example_frif.pdf",
         bedFile = c("promoter", "promoter_flanking", "exon",
                     "intron", "utr3", "utr5"))
}
\keyword{BED}
\keyword{FRiF}
\keyword{FRiP}
