% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppro.R
\name{plotTSS}
\alias{plotTSS}
\title{Plot TSS enrichment}
\usage{
plotTSS(TSSfile)
}
\arguments{
\item{TSSfile}{TSS enrichment file}
}
\description{
This function plots the global TSS enrichment and produces pdf/png files.
}
\examples{
data("tss")
plotTSS(TSSfile = "TSSfile")
}
\keyword{TSS}
\keyword{enrichment}
