% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotTSSscores}
\alias{plotTSSscores}
\title{This function is meant to plot TSS scores.}
\usage{
plotTSSscores(prj, stats, cutoff = 6)
}
\arguments{
\item{prj}{A PEPr Project object}

\item{stats}{A looper derived stats_summary.tsv file}

\item{cutoff}{A TSS enrichment cutoff score for low quality scores}
}
\description{
This function is meant to plot TSS scores.
}
\keyword{TSS}
\keyword{scores}
