% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{readPepatacPeakCounts}
\alias{readPepatacPeakCounts}
\title{Read PEPATAC peak coverage files}
\usage{
readPepatacPeakCounts(prj)
}
\arguments{
\item{pep}{A PEPr project object}
}
\description{
Read PEPATAC peak coverage files
}
\keyword{coverage}
\keyword{peak}
\keyword{project}
