% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{reducePeaks}
\alias{reducePeaks}
\title{This function is meant to keep only the most significant non-overlapping
peaks.  It also trims peaks extending beyond the bounds of the chromosome.}
\usage{
reducePeaks(input, sample_name, chr_sizes, output = NA, normalize = FALSE)
}
\arguments{
\item{input}{Path to narrowPeak file}

\item{sample_name}{Sample name character string}

\item{chr_sizes}{Genome chromosome sizes file. <Chr> <Size>}

\item{output}{Output file name.}

\item{normalize}{Remove overlaps and normalize the score.}
}
\description{
This function is meant to keep only the most significant non-overlapping
peaks.  It also trims peaks extending beyond the bounds of the chromosome.
}
\keyword{fixed}
\keyword{peaks}
\keyword{reduce}
