% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{roundUpNice}
\alias{roundUpNice}
\title{The function rounds the up to the nearest "nice" number.}
\usage{
roundUpNice(x, nice = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
\arguments{
\item{x}{A vector of numbers}

\item{nice}{Values to round up to}
}
\description{
From:
https://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
}
