% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{sampleName}
\alias{sampleName}
\title{Derive the sample name from input file and return with full path}
\usage{
sampleName(path, num_fields = 2, delim = "_")
}
\arguments{
\item{path}{A path to a file for which you wish to extract the sample name}

\item{num_fields}{An integer representing the number of fields to strip}

\item{delim}{A delimiter for the fields splitting a path or string}
}
\description{
Derive the sample name from input file and return with full path
}
