% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{setPanelSize}
\alias{setPanelSize}
\title{Set the panel width/height of a ggplot to a fixed value.}
\usage{
setPanelSize(
  p = NULL,
  g = ggplotGrob(p),
  file = NULL,
  margin = unit(1, "in"),
  width = unit(4, "in"),
  height = unit(4, "in")
)
}
\arguments{
\item{p}{a ggplot object}

\item{g}{a ggplot grob}

\item{file}{filepath}

\item{margin}{plot margins}

\item{width}{plot width}

\item{height}{plot height}
}
\description{
From:
https://github.com/baptiste/egg/blob/master/R/setPanelSize.r
}
