% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{splitDataTable}
\alias{splitDataTable}
\title{Efficiently split a data.table by a column in the table}
\usage{
splitDataTable(DT, split_factor)
}
\arguments{
\item{DT}{Data.table to split}

\item{split_factor}{Column to split, which can be a character vector
or an integer.}
}
\value{
List of data.table objects, split by column
}
\description{
Efficiently split a data.table by a column in the table
}
\examples{
  DT = data.table(letters, grp = rep(c("group1", "group2"), 13))
  splitDataTable(DT, "grp")
  splitDataTable(DT, 2)
}
