% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{summarizer}
\alias{summarizer}
\title{This function is meant to plot multiple summary graphs from the summary table 
made by the Looper summarize command}
\usage{
summarizer(project, output_dir)
}
\arguments{
\item{project}{A PEPr Project object}

\item{output_dir}{A PEP project output directory path}
}
\description{
This function is meant to plot multiple summary graphs from the summary table 
made by the Looper summarize command
}
\keyword{PEPATAC}
\keyword{summarize}
