% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{theme_PEPATAC}
\alias{theme_PEPATAC}
\title{A standardized ggplot theme for PEPATAC plots}
\usage{
theme_PEPATAC(base_family = "sans", ...)
}
\description{
A standardized ggplot theme for PEPATAC plots
}
\examples{
theme_PEPATAC()
}
\keyword{ggplot2}
\keyword{theme}
