% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{yamlToDT}
\alias{yamlToDT}
\title{Internal helper function for \code{summarizer}
Convert the new pipestat yaml outputs into R readable data.table file}
\usage{
yamlToDT(sample_name, yaml_file)
}
\arguments{
\item{sample_name}{A list project sample names}

\item{yaml_file}{A project level stats summary yaml file}
}
\description{
Internal helper function for \code{summarizer}
Convert the new pipestat yaml outputs into R readable data.table file
}
