% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{calcFRiF}
\alias{calcFRiF}
\title{Calculate the Fraction of Reads in Features (FRiF)}
\usage{
calcFRiF(bedFile, total, reads)
}
\arguments{
\item{bedFile}{A BED format file}

\item{total}{Number of aligned reads (or number of aligned bases)}

\item{reads}{If TRUE, we're working with read counts.
If FALSE, we're working with absolute number of bases}
}
\description{
This function calculates the fraction of reads in a feature and returns
a modified BED file with the cumulative sum of reads, cumulative size
of covered features, the fraction of reads in those features, and the
number of total features.
}
\examples{
calcFRiF()
}
\keyword{FRiF}
