% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{consensusPeaks}
\alias{consensusPeaks}
\title{This function is meant to identify a project level set of consensus peaks.}
\usage{
consensusPeaks(project, output_dir, results_subdir, assets)
}
\arguments{
\item{project}{A PEPr Project object}

\item{output_dir}{A PEP project output directory path}

\item{results_subdir}{A PEP project results subdirectory path}

\item{assets}{A data.table containing file assets}
}
\description{
This function is meant to identify a project level set of consensus peaks.
}
\keyword{consensus}
\keyword{peaks}
