% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{consensusPeaks}
\alias{consensusPeaks}
\title{This function is meant to identify a project level set of consensus peaks.}
\usage{
consensusPeaks(
  sample_table,
  summary_dir,
  results_subdir,
  assets,
  min_samples = 2,
  min_score = 5,
  min_olap = 1
)
}
\arguments{
\item{sample_table}{A data.table containing sample names and corresponding
genomes.}

\item{summary_dir}{A directory path to place results of this analysis}

\item{results_subdir}{A project results subdirectory path}

\item{assets}{A data.table containing file assets}

\item{min_samples}{A minimum number of samples a peak must be present
in to keep.}

\item{min_score}{A minimum peak score to keep an individual peak.}

\item{min_olap}{A minimum number of bases between peaks to be 
considered overlapping.}
}
\description{
This function is meant to identify a project level set of consensus peaks.
}
\keyword{consensus}
\keyword{peaks}
