% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{createAssetsSummary}
\alias{createAssetsSummary}
\title{Create and return assets summary data table}
\usage{
createAssetsSummary(samples, results_subdir)
}
\arguments{
\item{samples}{A PEP project character vector of sample names}

\item{results_subdir}{A PEP project results subdirectory path}
}
\description{
Create and return assets summary data table
}
