% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{createAssetsSummary}
\alias{createAssetsSummary}
\title{Create and return assets spreadsheet and save the spreadsheet to file}
\usage{
createAssetsSummary(project, output_dir, results_subdir)
}
\arguments{
\item{project}{A PEPr project object}

\item{output_dir}{A PEP project output directory path}

\item{results_subdir}{A PEP project results subdirectory path}
}
\description{
Create and return assets spreadsheet and save the spreadsheet to file
}
