% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{peakCounts}
\alias{peakCounts}
\title{Produce a project level peak counts table}
\usage{
peakCounts(
  sample_table,
  summary_dir,
  results_subdir,
  assets,
  poverlap = FALSE,
  norm = FALSE,
  cutoff = 2
)
}
\arguments{
\item{sample_table}{A data.table containing sample names and their 
corresponding genome}

\item{summary_dir}{A PEP project summary directory path}

\item{results_subdir}{A PEP project results subdirectory path}

\item{assets}{A data.table containing file assets}

\item{poverlap}{Weight counts by the percentage overlap with peak}

\item{norm}{Use normalized read counts}

\item{cutoff}{Only keep peaks present in the `cutoff` number of samples}
}
\description{
Produce a project level peak counts table
}
\keyword{counts}
\keyword{peak}
\keyword{project}
