% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{peakCounts}
\alias{peakCounts}
\title{Produce a project level peak counts table}
\usage{
peakCounts(project, output_dir, results_subdir, assets)
}
\arguments{
\item{project}{A PEPr project object}

\item{output_dir}{A PEP project output directory path}

\item{results_subdir}{A PEP project results subdirectory path}

\item{assets}{A data.table containing file assets}
}
\description{
Produce a project level peak counts table
}
\keyword{counts}
\keyword{peak}
\keyword{project}
