% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{peakCounts}
\alias{peakCounts}
\title{Produce a project level peak counts table}
\usage{
peakCounts(pep)
}
\arguments{
\item{pep}{A PEP configuration file}
}
\description{
Produce a project level peak counts table
}
\keyword{counts}
\keyword{peak}
\keyword{project}
