% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotAlignedRaw}
\alias{plotAlignedRaw}
\title{This function is meant to plot project level raw aligned reads.}
\usage{
plotAlignedRaw(prj, stats)
}
\arguments{
\item{prj}{A PEPr Project object}

\item{stats}{A looper derived stats_summary.tsv file}
}
\description{
This function is meant to plot project level raw aligned reads.
}
\keyword{aligned}
\keyword{raw}
\keyword{reads}
