% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotFRiF}
\alias{plotFRiF}
\title{Calculate and return a plot of either the expected or cumulative 
distribution of regions in a provided list of genomic partitions.}
\usage{
plotFRiF(
  query,
  partition_list,
  feature_names = NULL,
  type = c("expected", "cumulative")
)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions.}

\item{partition_list}{An ORDERED and NAMED list of genomic partitions
GRanges. This list must be in priority order; the input will be assigned
to the first partition it overlaps.}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the partitionList object.}

\item{type}{Plot either the 'expected' or 'cumulative' distribution.}
}
\description{
Calculate and return a plot of either the expected or cumulative 
distribution of regions in a provided list of genomic partitions.
}
