% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotLibSizes}
\alias{plotLibSizes}
\title{This function is meant to plot library sizes.}
\usage{
plotLibSizes(prj, stats)
}
\arguments{
\item{prj}{A PEPr Project object}

\item{stats}{A looper derived stats_summary.tsv file}
}
\description{
This function is meant to plot library sizes.
}
\keyword{library}
\keyword{size}
