% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{plotTSS}
\alias{plotTSS}
\title{Plot TSS enrichment}
\usage{
plotTSS(TSSfile, cutoff = 6)
}
\arguments{
\item{TSSfile}{TSS enrichment file}

\item{cutoff}{A TSS enrichment score cutoff value for low quality samples}
}
\description{
This function plots the global TSS enrichment.
}
\examples{
data("tss")
plotTSS(TSSfile = "tss")
}
\keyword{TSS}
\keyword{enrichment}
