% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPATACr.R
\name{readPepatacPeakCounts}
\alias{readPepatacPeakCounts}
\title{Read PEPATAC peak coverage files}
\usage{
readPepatacPeakCounts(prj, results_subdir)
}
\arguments{
\item{prj}{A PEPr project object}

\item{results_subdir}{A PEP project results subdirectory path}
}
\description{
Read PEPATAC peak coverage files
}
\keyword{coverage}
\keyword{peak}
\keyword{project}
